backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-chess==1.999
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.8.1.78  # для компьютерного зрения (опционально)
numpy==1.24.3
